fixed_record

Module prelude

Re-exports

pub use crate::RecordWithList;

pub use crate::fixed_record;

pub use crate::ByteRangeBounds;

pub use crate::Error;

pub use crate::Fixed;

pub use crate::FixedRecord;

pub use crate::Reader;

pub use crate::RecordSeparator;

pub use crate::Writer;

Traits

FromBytes

Types for which any bit pattern is valid.

Immutable

Types which are free from interior mutability.

IntoBytes

Types that can be converted to an immutable slice of initialized bytes.

KnownLayout

Indicates that zerocopy can reason about certain aspects of a type’s layout.

Attribute Macros

fixed_record

Generates helper types and implementations for fixed-width records. 固定長レコード用の補助型と実装を生成する attribute macro です。

Derive Macros

FromBytes

Immutable

IntoBytes

KnownLayout